+++ cef-20-demo-hub/src/main/resources/templates/sc-request.html
@@ -5,13 +5,13 @@
     <div th:include="fragments/default :: scripts" th:remove="tag"></div>
     <script src="../webjars/jquery-cookie/1.4.1-1/jquery.cookie.js"></script>
     <script src="../webjars/bootstrap-select/1.13.8/js/bootstrap-select.min.js"></script>
-    <script src="../js/sc-main.js"></script>
+    <script src="../js/sp-request.js"></script>
     <link href="../css/flatly-bootstrap.min.css" rel="stylesheet"/>
     <link rel="stylesheet" href="../webjars/bootstrap-select/1.13.8/css/bootstrap-select.min.css">
     <link rel="stylesheet" href="../css/sc-main.css">
     <link rel="stylesheet" href="../css/sc-cef-sp.css">
     <div th:include="fragments/default :: favicon" th:remove="tag"></div>
-    <title>CEF eIDAS 2.0.0 - Demo application hub</title>
+    <title>CEF eIDAS 2.2.0 - Test - SP</title>
 </head>
 <body>
